Update tests to support Xdelta3 version 3.0.11 In Xdelta3 version 3.0.11, well I think it was actually as of version 3.0.8 , the generated diff files are different. The result was that the tests which come with node-xdelta3 were no longer passing if you use a newer version of Xdelta. To fix this I'm updating the 'golden' diff txt1gent to one generated by Xdelta 3.0.11.
To ensure that the change is backwards compatible I've saved the 3.0.6 diff as
txt1gent.3.0.6 and added a new test, test_patch_basic_xdelta_3_0_6.js, which
if your using Xdelta3.0.11 will use the old golden diff (txt1gent.3.0.6) and
apply it to ensure it still works. If your using Xdelta 3.0.6 then it does
the same as test_patch_basic_1.js.

I've also updated test_diff_basic_1.js to try both the 3.0.6 and 3.0.11 versions
of the diff before failing so it should work with both version of Xdelta3.